+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The main focus is on user-observable behavior of the engine.
 * Implemented the [Intl.DateTimeFormat dateStyle & timeStyle](https://github.com/tc39/proposal-intl-datetime-style) proposal.
 
 ## Version 20.1.0
+* Implemented support for `export * as ns from "mod"`. It is available in ECMAScript 2020 mode (`--js.ecmascript-version=2020`).
 * Implemented the [class fields](https://github.com/tc39/proposal-class-fields) proposal for public and private instance fields. This feature is available by default in Node.js and can be enabled using the experimental option `js.class-fields`.
 * Added option `js.global-arguments` (default: true) for the non-standard `arguments` global property (disabled in Node.js).