+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The main focus is on user-observable behavior of the engine.
 * Added option `js.global-arguments` (default: true) for the non-standard `arguments` global property (disabled in Node.js).
 
 ## Version 20.0.0
-* Updated Node.js to version 12.14.0.
+* Updated Node.js to version 12.15.0.
 * Added `name` property to anonymous functions.
 * `String.prototype.matchAll` is available in ECMAScript 2020 mode (`--js.ecmascript-version=2020`) only.
 * Implemented the [String.prototype.replaceAll](https://github.com/tc39/proposal-string-replaceall) proposal. It is available in ECMAScript 2020 mode (`--js.ecmascript-version=2020`).